elsai-stt/pyproject.toml
[project]
name = "elsai-stt"
version = "0.1.0"
description = ""
authors = [
    {name = "abhinivesh0305",email = "abhinivesh.s@optisolbusiness.com"}
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "audioop-lts (>=0.2.1,<0.3.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.84.0,<2.0.0)",
    "numpy (>=2.2.6,<3.0.0)"

 
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
